requirements.txt
# Core dependencies for yellow.py
numpy>=1.21.0
Pillow>=8.0.0

# Dependencies for batch processing wrapper
selenium>=4.0.0
webdriver-manager>=3.8.0

# Optional: for better performance
psutil>=5.8.0

# Development/testing
pytest>=6.0.0
